Remove unnecessary _super call
This call is not needed, the route extends Ember.Route, this PR is in response to someone who it confused on slack.

Either way, it's definitely wrong. You can't just call `this._super()` in beforeModel and expect it to work. If you actually have a superclass hook to call, you need to take promises into account